Relocate the repositories
