+++ sorellaFront/src/app/app.module.ts
@@ -9,7 +9,7 @@ import { AppComponent } from './app.component';
 import { AbstractProductService } from 'src/application/abstractions';
 import { ProductService } from 'src/application/services';
 import { AbstractProductRepository } from 'src/domain/repositories';
-import { ProductImplementationRepository } from 'src/application/repositories';
+import { ProductImplementationRepository } from 'src/infrastructure/repositories';
 import { HttpClientModule } from '@angular/common/http';
